requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.16.0
crispy-bootstrap4==2024.1
cryptography==42.0.7
Django==5.0.6
django-appconf==1.0.6
django-bootstrap-datepicker-plus==5.0.5
django-crispy-forms==2.1
django-select2==8.1.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
pyDes==2.0.1
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
